Fixed list, expanded README and added systemd
Fixed the bug with displaying a listing of all accounts

Added some docs to the README.md for the oohhc-acctd service with
a list of parameters to run either on the command line or using
a systemd service.  I also put in the defaults

Added the oohhc-acctd.service file with the same directory
structure used in the other FSAAAS repos.